pyproject.toml
[tool.poetry]
name = "test-alteasy"
version = "0.1.0"
description = ""
authors = ["Kiryl Shvedau <80332673+Svensson17@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
